Added source path to script and style destinations
The current script and style copying ignores the file path of the source.
Copying multiple files with the same name, but different paths, will fail as
the files will be overridden and only the last of the duplicate files will
be available after the copy. This is a problem when depending on bower
components as most of these use the index.js file name.

Preserving source path when copying scripts and styles to prevent the issue.